+++ src/components/ClickerButton/ClickerButton2.tsx
